requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==0.9.0
    # via
    #   -r requirements.in
    #   tensorboard
appnope==0.1.4
    # via
    #   ipykernel
    #   ipython
argon2-cffi==23.1.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
attrs==19.3.0
    # via
    #   -r requirements.in
    #   jsonschema
backcall==0.1.0
    # via
    #   -r requirements.in
    #   ipython
bleach==3.3.0
    # via
    #   -r requirements.in
    #   nbconvert
cachetools==4.0.0
    # via
    #   -r requirements.in
    #   google-auth
certifi==2019.11.28
    # via
    #   -r requirements.in
    #   requests
cffi==1.17.1
    # via argon2-cffi-bindings
chardet==3.0.4
    # via
    #   -r requirements.in
    #   requests
cycler==0.10.0
    # via
    #   -r requirements.in
    #   matplotlib
decorator==4.4.2
    # via
    #   -r requirements.in
    #   ipython
    #   traitlets
defusedxml==0.6.0
    # via
    #   -r requirements.in
    #   nbconvert
entrypoints==0.3
    # via
    #   -r requirements.in
    #   nbconvert
filelock==3.16.1
    # via torch
fsspec==2024.9.0
    # via torch
future==0.18.2
    # via
    #   -r requirements.in
    #   pytorch-lightning
google-auth==1.35.0
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.1
    # via
    #   -r requirements.in
    #   tensorboard
grpcio==1.66.2
    # via
    #   -r requirements.in
    #   tensorboard
idna==2.9
    # via
    #   -r requirements.in
    #   requests
importlib-metadata==1.5.0
    # via -r requirements.in
ipykernel==5.1.4
    # via
    #   -r requirements.in
    #   ipywidgets
    #   notebook
ipython==7.13.0
    # via
    #   -r requirements.in
    #   ipykernel
    #   ipywidgets
ipython-genutils==0.2.0
    # via
    #   -r requirements.in
    #   nbformat
    #   notebook
    #   traitlets
ipywidgets==7.5.1
    # via -r requirements.in
jedi==0.16.0
    # via
    #   -r requirements.in
    #   ipython
jinja2==2.11.3
    # via
    #   -r requirements.in
    #   nbconvert
    #   notebook
    #   torch
jsonschema==3.2.0
    # via
    #   -r requirements.in
    #   nbformat
jupyter-client==6.0.0
    # via
    #   -r requirements.in
    #   ipykernel
    #   notebook
jupyter-core==4.6.3
    # via
    #   -r requirements.in
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
kiwisolver==1.1.0
    # via
    #   -r requirements.in
    #   matplotlib
markdown==3.2.1
    # via
    #   -r requirements.in
    #   tensorboard
markupsafe==1.1.1
    # via
    #   -r requirements.in
    #   jinja2
matplotlib==3.2.0
    # via -r requirements.in
mistune==0.8.4
    # via
    #   -r requirements.in
    #   nbconvert
mpmath==1.3.0
    # via sympy
nbconvert==5.6.1
    # via
    #   -r requirements.in
    #   notebook
nbformat==5.0.4
    # via
    #   -r requirements.in
    #   ipywidgets
    #   nbconvert
    #   notebook
networkx==3.2.1
    # via torch
notebook==6.1.5
    # via
    #   -r requirements.in
    #   widgetsnbextension
numpy==2.0.2
    # via
    #   -r requirements.in
    #   matplotlib
    #   opencv-python
    #   pandas
    #   pytorch-lightning
    #   scipy
    #   tensorboard
    #   torchvision
oauthlib==3.1.0
    # via
    #   -r requirements.in
    #   requests-oauthlib
opencv-python==4.5.5.64
    # via -r requirements.in
packaging==24.1
    # via bleach
pandas==2.2.3
    # via -r requirements.in
pandocfilters==1.4.2
    # via
    #   -r requirements.in
    #   nbconvert
parso==0.6.2
    # via
    #   -r requirements.in
    #   jedi
pexpect==4.8.0
    # via
    #   -r requirements.in
    #   ipython
pickleshare==0.7.5
    # via
    #   -r requirements.in
    #   ipython
pillow==8.1.1
    # via
    #   -r requirements.in
    #   torchvision
prometheus-client==0.7.1
    # via
    #   -r requirements.in
    #   notebook
prompt-toolkit==3.0.4
    # via
    #   -r requirements.in
    #   ipython
protobuf==3.11.3
    # via
    #   -r requirements.in
    #   tensorboard
ptyprocess==0.6.0
    # via
    #   -r requirements.in
    #   pexpect
    #   terminado
pyasn1==0.4.8
    # via
    #   -r requirements.in
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   -r requirements.in
    #   google-auth
pycparser==2.22
    # via cffi
pygments==2.7.4
    # via
    #   -r requirements.in
    #   ipython
    #   nbconvert
pyparsing==2.4.6
    # via
    #   -r requirements.in
    #   matplotlib
pyrsistent==0.15.7
    # via
    #   -r requirements.in
    #   jsonschema
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   jupyter-client
    #   matplotlib
    #   pandas
pytorch-lightning==0.7.1
    # via -r requirements.in
pytz==2024.2
    # via
    #   -r requirements.in
    #   pandas
pyzmq==19.0.0
    # via
    #   -r requirements.in
    #   jupyter-client
    #   notebook
requests==2.23.0
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.0
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
rsa==4.7
    # via
    #   -r requirements.in
    #   google-auth
scipy==1.13.1
    # via -r requirements.in
send2trash==1.5.0
    # via
    #   -r requirements.in
    #   notebook
six==1.14.0
    # via
    #   -r requirements.in
    #   absl-py
    #   bleach
    #   cycler
    #   google-auth
    #   grpcio
    #   jsonschema
    #   protobuf
    #   pyrsistent
    #   python-dateutil
    #   tensorboard
    #   traitlets
sympy==1.13.3
    # via torch
tensorboard==2.1.1
    # via
    #   -r requirements.in
    #   pytorch-lightning
terminado==0.8.3
    # via
    #   -r requirements.in
    #   notebook
testpath==0.4.4
    # via
    #   -r requirements.in
    #   nbconvert
torch==2.2.2
    # via
    #   -r requirements.in
    #   pytorch-lightning
    #   torchvision
torchvision==0.17.2
    # via -r requirements.in
tornado==6.0.4
    # via
    #   -r requirements.in
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tqdm==4.43.0
    # via
    #   -r requirements.in
    #   pytorch-lightning
traitlets==4.3.3
    # via
    #   -r requirements.in
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   nbconvert
    #   nbformat
    #   notebook
typing-extensions==4.12.2
    # via torch
tzdata==2024.2
    # via pandas
urllib3==1.25.8
    # via
    #   -r requirements.in
    #   requests
wcwidth==0.1.8
    # via
    #   -r requirements.in
    #   prompt-toolkit
webencodings==0.5.1
    # via
    #   -r requirements.in
    #   bleach
werkzeug==1.0.0
    # via
    #   -r requirements.in
    #   tensorboard
wheel==0.44.0
    # via tensorboard
widgetsnbextension==3.5.1
    # via
    #   -r requirements.in
    #   ipywidgets
zipp==3.1.0
    # via
    #   -r requirements.in
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
